# feat: async build word index map when init lsp state